Fixed some Win-related issues in Charmm and in the project files

+++ include/BALL/MOLMEC/CHARMM/charmmTorsion.h
@@ -1,7 +1,7 @@
 // -*- Mode: C++; tab-width: 2; -*-
 // vi: set ts=2:
 //
-// $Id: charmmTorsion.h,v 1.15 2003/04/29 16:07:16 oliver Exp $
+// $Id: charmmTorsion.h,v 1.16 2003/04/30 06:05:18 oliver Exp $
 // Molecular Mechanics: CHARMM force field, proper torsion component
 
 #ifndef BALL_MOLMEC_CHARMM_CHARMMTORSION_H
@@ -71,7 +71,7 @@ namespace BALL
 				f = (unsigned char)t.values.f;
 
 				// convert phase from degrees to radiant
-				phase = (BALL::Constants::PI / 180.0) * t.values.phase;
+				phase = (float)(BALL::Constants::PI / 180.0) * t.values.phase;
 			}
 		};
